las explosiones ya funcionan mas o menos

+++ skeleton/ForceGenerator.h
@@ -12,7 +12,7 @@ public:
 	}
 
 	virtual void applyForce(Particle* particle) = 0;
-	virtual void update(float deltaTime) = 0;
+	virtual void update(double deltaTime) = 0;
 
 };
